refactor: code refactorings
