+++ src/CreditCardChallenge.Web/Pages/Index.razor.css
@@ -95,3 +95,15 @@ input {
 .rotate {
     transform: rotateY(180deg);
 }
+
+/* Chrome, Safari, Edge, Opera */
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+}
+
+/* Firefox */
+input[type=number] {
+    -moz-appearance: textfield;
+}
