+++ src/CreditCardChallenge.Web/Program.cs
@@ -1,4 +1,6 @@
 using CreditCardChallenge.Web;
+using CreditCardChallenge.Web.Models;
+using FluentValidation;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 
@@ -7,5 +9,6 @@ builder.RootComponents.Add<App>("#app");
 builder.RootComponents.Add<HeadOutlet>("head::after");
 
 builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
+builder.Services.AddValidatorsFromAssemblyContaining<CardDetailsValidator>(ServiceLifetime.Transient);
 
 await builder.Build().RunAsync();
